WIP: Add to listener on the SMS received event
It is not yet complete,
some of the code should already handle the subsciption to the event but
the function that should be triggered by this event does not fire.

I should look into the priority, of maybe the event is aborted by an app
that has a higher priority.